agent/requirements.txt
gradient-adk>=0.1.0
langgraph>=0.2.0
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
